+++ lib/rubygems/tasks/console.rb
@@ -87,13 +87,8 @@ module Gem
         arguments.push(*@options)
 
         if @project.bundler?
-          if (@command == DEFAULT_CONSOLE && @options.empty?)
-            # use `bundle console` unless were were using custom command/options
-            arguments = ['bundle', 'console']
-          else
-            # run under `bundle exec`
-            arguments.unshift('bundle', 'exec')
-          end
+          # run under `bundle exec`
+          arguments.unshift('bundle', 'exec')
         end
 
         return run(*arguments)